+++ server/impl/src/main/java/com/walmartlabs/concord/server/console/CustomFormServiceImpl.java
@@ -10,6 +10,8 @@ import com.walmartlabs.concord.server.api.process.ProcessStatus;
 import com.walmartlabs.concord.server.cfg.FormServerConfiguration;
 import com.walmartlabs.concord.server.history.ProcessHistoryDao;
 import com.walmartlabs.concord.server.process.ConcordFormService;
+import com.walmartlabs.concord.server.process.FormUtils;
+import com.walmartlabs.concord.server.process.FormUtils.ValidationException;
 import com.walmartlabs.concord.server.process.PayloadManager;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.form.DefaultFormValidatorLocale;
@@ -17,7 +19,6 @@ import io.takari.bpm.form.Form;
 import io.takari.bpm.form.FormSubmitResult;
 import io.takari.bpm.form.FormSubmitResult.ValidationError;
 import io.takari.bpm.form.FormValidatorLocale;
-import io.takari.bpm.model.form.DefaultFormFields;
 import io.takari.bpm.model.form.FormDefinition;
 import io.takari.bpm.model.form.FormField;
 import io.takari.bpm.model.form.FormField.Cardinality;
@@ -38,7 +39,10 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @Named
 public class CustomFormServiceImpl implements CustomFormService, Resource {
@@ -123,7 +127,7 @@ public class CustomFormServiceImpl implements CustomFormService, Resource {
         try {
             Map<String, Object> m = new HashMap<>();
             try {
-                m = convert(form, convert(data));
+                m = FormUtils.convert(validatorLocale, form, FormUtils.convert(data));
 
                 FormSubmitResult r = formService.submit(processInstanceId, formInstanceId, m);
                 if (r.isValid()) {
@@ -165,7 +169,7 @@ public class CustomFormServiceImpl implements CustomFormService, Resource {
                     writeData(dst, prepareData(form, m, r.getErrors()));
                 }
             } catch (ValidationException e) {
-                ValidationError err = new ValidationError(e.field.getName(), e.message);
+                ValidationError err = new ValidationError(e.getField().getName(), e.getMessage());
                 FormData d = prepareData(form, m, Collections.singletonList(err));
                 writeData(dst, d);
             }
@@ -271,113 +275,6 @@ public class CustomFormServiceImpl implements CustomFormService, Resource {
         Path dst = baseDir.resolve("data.js");
         String s = String.format(DATA_FILE_TEMPLATE, objectMapper.writeValueAsString(data));
         Files.write(dst, s.getBytes(), StandardOpenOption.TRUNCATE_EXISTING);
-    }
-
-    private static FormField findField(FormDefinition fd, String fieldName) {
-        for (FormField f : fd.getFields()) {
-            if (fieldName.equals(f.getName())) {
-                return f;
-            }
-        }
-        return null;
-    }
-
-    private static Map<String, Object> convert(MultivaluedMap<String, String> data) {
-        Map<String, Object> m = new HashMap<>();
-        if (data != null) {
-            data.forEach((k, v) -> {
-                if (v == null) {
-                    return;
-                }
-
-                int size = v.size();
-                if (size == 0) {
-                    return;
-                } else if (size == 1) {
-                    m.put(k, v.get(0));
-                } else {
-                    m.put(k, v);
-                }
-            });
-        }
-        return m;
-    }
-
-    // TODO this probably should be a part of the bpm engine's FormService
-    private Map<String, Object> convert(Form form, Map<String, Object> m) throws ValidationException {
-        FormDefinition fd = form.getFormDefinition();
-
-        Map<String, Object> m2 = new HashMap<>();
-        for (Map.Entry<String, Object> e : m.entrySet()) {
-            String k = e.getKey();
-
-            FormField f = findField(fd, k);
-            if (f == null) {
-                continue;
-            }
-
-            Object v = convert(fd.getName(), f, null, e.getValue());
-            if (v == null) {
-                continue;
-            }
-
-            m2.put(k, v);
-        }
-        return m2;
-    }
-
-    private Object convert(String formName, FormField f, Integer idx, Object v) throws ValidationException {
-        if (v instanceof String) {
-            String s = (String) v;
-            if (s.isEmpty()) {
-                return null;
-            }
-
-            switch (f.getType()) {
-                case DefaultFormFields.IntegerField.TYPE: {
-                    try {
-                        return Long.parseLong(s);
-                    } catch (NumberFormatException e) {
-                        throw new ValidationException(f, s, validatorLocale.expectedInteger(formName, f.getName(), idx, s));
-                    }
-                }
-                case DefaultFormFields.DecimalField.TYPE: {
-                    try {
-                        return Double.parseDouble(s);
-                    } catch (NumberFormatException e) {
-                        throw new ValidationException(f, s, validatorLocale.expectedDecimal(formName, f.getName(), idx, s));
-                    }
-                }
-            }
-        } else if (v instanceof List) {
-            List<?> l = (List<?>) v;
-            if (l.isEmpty()) {
-                return null;
-            }
-
-            List<Object> ll = new ArrayList<>(l.size());
-            int i = 0;
-            for (Object o : l) {
-                ll.add(convert(formName, f, i, o));
-                i++;
-            }
-            return ll;
-        }
-
-        return v;
-    }
-
-    private static class ValidationException extends Exception {
-
-        private final FormField field;
-        private final String input;
-        private final String message;
-
-        private ValidationException(FormField field, String input, String message) {
-            this.field = field;
-            this.input = input;
-            this.message = message;
-        }
     }
 
     @JsonInclude(Include.NON_EMPTY)